Create basic-calculator.py
Initial commit of the basic calculator project I have created in Python. This first version of the program allows the user to select one of four arithmetic operations (addition, subtraction, multiplication, division) to perform on two numbers, then it prints the result. This first version of the program has been kept simple for future changes.